Break out root task creation to reduce file size
